+++ tests/test_address_validation_service.py
@@ -3,16 +3,18 @@ Test module for the Fedex AddressValidationService WSDL.
 """
 
 import unittest
-
+import logging
 import sys
 
 sys.path.insert(0, '..')
 from fedex.services.address_validation_service import FedexAddressValidationRequest
 
 # Common global config object for testing.
-from common import get_test_config
+from common import get_fedex_config
 
-CONFIG_OBJ = get_test_config()
+CONFIG_OBJ = get_fedex_config()
+
+logging.getLogger('suds').setLevel(logging.ERROR)
 
 
 @unittest.skipIf(not CONFIG_OBJ.account_number, "No credentials provided.")