+++ tests/test_ship_service.py
@@ -3,7 +3,7 @@ Test module for the Fedex ShipService WSDL.
 """
 
 import unittest
-
+import logging
 import sys
 
 sys.path.insert(0, '..')
@@ -11,9 +11,11 @@ from fedex.services.ship_service import FedexProcessShipmentRequest
 from fedex.services.ship_service import FedexDeleteShipmentRequest
 
 # Common global config object for testing.
-from common import get_test_config
+from common import get_fedex_config
 
-CONFIG_OBJ = get_test_config()
+CONFIG_OBJ = get_fedex_config()
+
+logging.getLogger('suds').setLevel(logging.ERROR)
 
 
 @unittest.skipIf(not CONFIG_OBJ.account_number, "No credentials provided.")